Moves incident dashboard components to the dashboard directory
