+++ src/dispatch/static/dispatch/src/dashboard/incident/IncidentCommandersLocationBarChartCard.vue
