+++ src/dispatch/static/dispatch/src/dashboard/incident/IncidentCostBarChartCard.vue
