+++ src/dispatch/static/dispatch/src/dashboard/incident/IncidentForecastCard.vue
