+++ src/dispatch/static/dispatch/src/dashboard/incident/IncidentHeatmapCard.vue
