+++ src/dispatch/static/dispatch/src/dashboard/incident/IncidentParticipantsTeamBarChartCard.vue
