+++ src/dispatch/static/dispatch/src/dashboard/incident/IncidentPriorityBarChartCard.vue
