+++ src/dispatch/static/dispatch/src/dashboard/incident/IncidentResolveTimeCard.vue
