+++ src/dispatch/static/dispatch/src/dashboard/incident/IncidentTypeBarChartCard.vue
